examples/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-adk>=0.4.2",
    "ag-ui-langgraph>=0.0.25",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "google-adk>=1.20.0",
    "langchain-litellm",
    "litellm>=1.81.5",
    "uvicorn>=0.40.0",
    "linkos",
    "langchain-mcp-adapters>=0.2.1",
]

[tool.uv.sources]
linkos = { path = "../linkos/python/", editable = true }
